requirements.txt
Django==3.1.7
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
phonenumbers==8.12.19
